+++ crates/homeboy-core/src/config/app.rs
@@ -4,20 +4,12 @@ use std::collections::HashMap;
 use super::{AppPaths, ConfigImportable, ConfigManager};
 use crate::Result;
 
-fn default_cli_path() -> String {
-    "wp".to_string()
-}
-
 fn default_database_host() -> String {
     "127.0.0.1".to_string()
 }
 
 fn default_local_db_port() -> u16 {
     33306
-}
-
-fn is_default_cli_path(v: &String) -> bool {
-    v == "wp"
 }
 
 fn is_default_database_host(v: &String) -> bool {
@@ -32,9 +24,6 @@ fn is_default_local_db_port(v: &u16) -> bool {
 #[serde(rename_all = "camelCase")]
 pub struct AppConfig {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub active_project_id: Option<String>,
-
-    #[serde(skip_serializing_if = "Option::is_none")]
     pub default_changelog_next_section_label: Option<String>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -43,11 +32,8 @@ pub struct AppConfig {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub installed_modules: Option<HashMap<String, InstalledModuleConfig>>,
 
-    #[serde(
-        default = "default_cli_path",
-        skip_serializing_if = "is_default_cli_path"
-    )]
-    pub default_cli_path: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub default_cli_path: Option<String>,
 
     #[serde(
         default = "default_database_host",
@@ -65,11 +51,10 @@ pub struct AppConfig {
 impl Default for AppConfig {
     fn default() -> Self {
         Self {
-            active_project_id: None,
             default_changelog_next_section_label: None,
             default_changelog_next_section_aliases: None,
             installed_modules: None,
-            default_cli_path: default_cli_path(),
+            default_cli_path: None,
             default_database_host: default_database_host(),
             default_local_db_port: default_local_db_port(),
         }